Added placement history for MasterworksFair Added weaponscore to MasterworksFairEntry Fixed reversed weaponscore sort Fixed database players not updating after getting rewards

+++ src/main/java/com/ebicep/warlords/pve/weapons/weaponaddons/WeaponScore.java
@@ -31,13 +31,13 @@ public interface WeaponScore {
         System.out.println("Max Score: " + maxScore);
     }
 
-    default double getWeaponScore() {
+    default float getWeaponScore() {
         List<Double> averageScores = getWeaponScoreAverageValues();
         double sum = 0;
         for (Double d : averageScores) {
             sum += d;
         }
-        return Math.round(sum / averageScores.size() * 10000) / 100.0;
+        return Math.round(sum / averageScores.size() * 10000) / 100f;
     }
 
     default String getWeaponScoreString() {
